requirements.txt
Flask==2.0.1
flask-cors==3.0.10
PyJWT==2.8.0
bcrypt==3.2.0